BUG#: 5947 TITLE: Pegasus support for 64-bit Windows platform
DESCRIPTION:

Add support for the WIN64_IA64_MSVC and WIN64_X86_64_MSVC platforms and remove warnings associated with 64-bit windows compilation.

+++ src/Clients/repupgrade/RepositoryUpgrade.cpp
@@ -2056,11 +2056,11 @@ void RepositoryUpgrade::_logRequestToFile(
 
             if( contentLength > 0 )
             {
-                contentOffset = data.size() - contentLength;
+                contentOffset = (Uint32)data.size() - contentLength;
             }
             else
             {
-                contentOffset = data.size();
+                contentOffset = (Uint32)data.size();
             }
 
             //
